Implement first converter
In order to implement the conversion in an extendable way I have
added a Converter interface for which different converters can be
implemented. A Converter implementations aims to convert a certain
type or a certain type set.

The first converter is used for simple types, respectively primitive
types, their corresponding wrapper classes and Strings.

The converters are ordered by priorities in order to establish a hierarchy:
ArrayList, List, Collection instead of Collection, List, ArrayList
concerning the conversion order.

Signed-off-by: Konrad Reiche <konrad.reiche@googlemail.com>