Add collection converter
For converting types which implement the Collection interface
I have added another converter. The converter simply iterates
the collection's element and applies the appropriate converter
on them.

In doing so I have also refactored the conversion interface
method to use a Property object as predicate instead of a
field. Not always the field itself is converted, but elements
of it as it is the case with collections.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>