Exclude empty objects
An object is empty if all its fields except the identifier
field are null. I have added a check for empty objects to
the object converter since they do not contain any relevant
information for the resource to be associated with.

Hence empty objects are not bound to the RDF model and not
associated with the resource in a RDF statement.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>