Remove unnecessary bind invocation
In the collection converter for each element the bind method
was invoked. This is not the responsibility of the collection
converter, but the responsibility of the corresponding element
converter.

Removing the method invocation has the effect that collections
of simple types like primitive types and Strings can now be
converted as well.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>