Add Enum converter
Until now enumerations were treated like Strings. However, with
the goal to provide a @SameAs annotation in the future it is
going to be more useful to process them differently. For this
an Enum converter was added which creates a new resource from
the Enum.

For the identifying URI the Enum constant value is used unless a
@RdfIdentifier annotation is provided.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>