Fix initial resource creation
When an object is newly bound to the RDF model a new
resource is created. In order to type that resource
the property rdfs:type is used. The type itself is
another resource with a corresponding URI identifier.

I made a mistake by not creating the type resource,
but just using the URI identifier as literal property.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>