Avoid field bindings of Enum base class
Since the Enum binding is handled by the general bind method it
had to be assured that fields of the Enum base class like ordinal
and name are not bound to the model. These fields are only important
for implementation detail but not on the abstraction level of modelling.

In this commit I have also added to the resource creation the case
when the URI ends with a hash #. If this is the case not a slash / will
be inserted between namespace and identifier but the identifier will be
appended directly after the namespace ending with the hash #.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>