Refactor conversion procedure using the Chain of Responsibility pattern
In the search for the pattern name the conversion procedure uses I came
over the Chain of Responsibility pattern. Even though this pattern is
normally used to process GUI events I would advocate to use it for a Java
object model serialization as well.

Hence, instead of deciding which pattern the current conversion procedure
uses I have decided to refactor the whole procedure applying the Chain of
Responsibility pattern.

The chain is formed by the different converters where each can have up to
one follow-up converter. If the converter cannot convert the given object
it passes the conversion invocation to its successor. The chain starts
with the Null Converter.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>