Remove identifier field filtering
Fields declared as identifier were skipped when binding the state of the
Java object to the RDF model. Often, however, this identifier is a
natural source, for instance a name attribute which should become a
property in the model. Hence, I have removed the condition where
identifier fields are skipped.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>